Move order direction enum to core module
