+++ saleor/graphql/core/enums.py
@@ -20,6 +20,21 @@ from .utils import str_to_enum
 #  to taxes
 
 
+class OrderDirection(graphene.Enum):
+    ASC = ""
+    DESC = "-"
+
+    @property
+    def description(self):
+        # Disable all the no-member violations in this function
+        # pylint: disable=no-member
+        if self == OrderDirection.ASC:
+            return "Specifies an ascending sort order."
+        if self == OrderDirection.DESC:
+            return "Specifies a descending sort order."
+        raise ValueError("Unsupported enum value: %s" % self.value)
+
+
 class ReportingPeriod(graphene.Enum):
     TODAY = "TODAY"
     THIS_MONTH = "THIS_MONTH"
