+++ saleor/graphql/product/enums.py
@@ -48,19 +48,6 @@ class ProductOrderField(graphene.Enum):
         raise ValueError("Unsupported enum value: %s" % self.value)
 
 
-class OrderDirection(graphene.Enum):
-    ASC = ""
-    DESC = "-"
-
-    @property
-    def description(self):
-        if self == OrderDirection.ASC:
-            return "Specifies an ascending sort order."
-        if self == OrderDirection.DESC:
-            return "Specifies a descending sort order."
-        raise ValueError("Unsupported enum value: %s" % self.value)
-
-
 class CollectionPublished(graphene.Enum):
     PUBLISHED = "published"
     HIDDEN = "hidden"
