+++ saleor/graphql/product/types/attributes.py
@@ -7,18 +7,14 @@ from graphene import relay
 from ....product import models
 from ....product.utils.attributes import AttributeAssignmentType
 from ...core.connection import CountableDjangoObjectType
+from ...core.enums import OrderDirection
 from ...core.resolvers import resolve_meta, resolve_private_meta
 from ...core.types import MetadataObjectType
 from ...decorators import permission_required
 from ...translations.fields import TranslationField
 from ...translations.types import AttributeTranslation, AttributeValueTranslation
 from ..descriptions import AttributeDescriptions, AttributeValueDescriptions
-from ..enums import (
-    AttributeInputTypeEnum,
-    AttributeSortField,
-    AttributeValueType,
-    OrderDirection,
-)
+from ..enums import AttributeInputTypeEnum, AttributeSortField, AttributeValueType
 
 COLOR_PATTERN = r"^(#[0-9a-fA-F]{3}|#(?:[0-9a-fA-F]{2}){2,4}|(rgb|hsl)a?\((-?\d+%?[,\s]+){2,3}\s*[\d\.]+%?\))$"  # noqa
 color_pattern = re.compile(COLOR_PATTERN)
